refactor: consolidate references/cpp/src/ into single file
Merge output.{cc,h}, utils.{cc,h}, reference.cc, and readme.cc into a
single main.cpp that can be compiled directly:

  c++ $(PKG_CONFIG_PATH="$out/lib/pkgconfig" pkg-config --cflags --libs milo_parser) references/cpp/src/main.cc -o cpp_test && ./cpp_test

This is useful to test the installation in a sandbox without pulling
lots of dependencies.
