--- references/cpp/src/readme.cc
@@ -1,42 +1,0 @@
-#include "milo.h"
-#include "stdio.h"
-#include "string.h"
-#include <cinttypes>
-
-int main() {
-  // Create the parser.
-  milo_parser::Parser* parser = milo_parser::milo_create();
-
-  // Prepare a message to parse.
-  const char* message = "HTTP/1.1 200 OK\r\nContent-Length: 3\r\n\r\nabc";
-
-  parser->context = (char*) message;
-
-  /*
-    Milo works using callbacks.
-
-    All callbacks have the same signature, which characterizes the payload:
-
-      * p: The current parser.
-      * at: The payload offset.
-      * len: The payload length.
-
-    The payload parameters above are relative to the last data sent to the milo_parse method.
-
-    If the current callback has no payload, both values are set to 0.
-  */
-  parser->callbacks.on_data = [](milo_parser::Parser* p, uintptr_t from, uintptr_t size) {
-    char* payload = reinterpret_cast<char*>(malloc(sizeof(char) * size));
-    strncpy(payload, reinterpret_cast<const char*>(p->context) + from, size);
-
-    printf("Pos=%" PRIuPTR " Body: %.*s\n", from, static_cast<int>(size), payload);
-    free(payload);
-  };
-  parser->active_callbacks |= milo_parser::CALLBACK_ACTIVE_ON_DATA;
-
-  // Now perform the main parsing using milo.parse. The method returns the number of consumed characters.
-  milo_parser::milo_parse(parser, reinterpret_cast<const unsigned char*>(message), strlen(message));
-
-  // Cleanup used resources.
-  milo_parser::milo_destroy(parser);
-}
